\name{fGarch-package}


\alias{fGarch-package}


\alias{fGarch}


\docType{package}


\title{GARCH Modelling Package}


\description{

    Package of econometric functions for modelling GARCH processes.
}


\details{

    \tabular{rl}{
        Package:    \tab fGarch\cr
        Type:       \tab Package\cr
        Version:    \tab 261.73.1\cr
        Date:       \tab 2008\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2008 Diethelm Wuertz and Rmetrics Foundation\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }

}


\section{Time Series Simulation}{

    contains functions to simulate artificial GARCH and APARCH
    time series processes.
    
}


\section{Parameter Estimation}{

    contains functions to fit the parameters of GARCH and APARCH 
    time series processes.
    
}


\section{Forecasting}{

    contains functions to forcecast mean and variance of GARCH and
    APARCH processes.
    
}


\section{Standardized Distribution Functions}{

    contains functions to model standardized distribution functions.
    
}


\section{OX Interface}{

    contains a Windows interface to OX.
    
}


\author{

    Diethelm Wuertz and Rmetrics Core Team.
    
}



\keyword{package}

