\name{fUGARCHSPEC-class}
\alias{fUGARCHSPEC-class}
\alias{.ugarchFit}
\alias{.ugarchSpec}

\docType{class}

\title{Class 'fUGARCHSPEC'}
\description{Class 'fUGARCHSPEC'.}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("fUGARCHSPEC", ...)}.

}
\section{Slots}{
  \describe{
    \item{\code{model}:}{Object of class \code{"list"} ~~ }
    \item{\code{distribution}:}{Object of class \code{"list"} ~~ }
    \item{\code{optimization}:}{Object of class \code{"list"} ~~ }
    \item{\code{documentation}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Methods}{
No methods defined with class \code{"fUGARCHSPEC"} in the signature.
}

\note{
  (GNB) This class seems to be meant for internal use by the package.
}

\seealso{
  class \code{"\linkS4class{fGARCH}"}
}

\examples{
showClass("fUGARCHSPEC")
}

\keyword{classes}
